// src/app/modules/Models/SchoolUser.ts
export class SchoolUser {
  first_name: string;
  last_name: string;
  middle_name: string;
  arabic_first_name: string;
  arabic_last_name: string;
  arabic_middle_name: string;
  email: string;
  date_of_birth: string;
  password: string;
  phone_number: string;
  mobile_number: string;
  place_of_birth: string;
  religion: string;
  is_active: boolean;
  address: string;
  city: string;
  building: string;
  floor: string;
  blood_type: string;
  user_gender: string;
  id: string;
  user_name: string;
  salt: string;
  created_at: string;
  updated_at: string;
  profile_pictureUrl: string;
  role_name: string;
}